refatorando validação & loading input

+++ server/src/app/controllers/ContactController.js
@@ -1,4 +1,5 @@
 const ContactsRepository = require("../repositories/ContactsRepository");
+const isValidUUID = require("../utils/isValidUUID");
 
 class ContactController {
  async index(req, res){
@@ -9,9 +10,14 @@ class ContactController {
  }
 
  async show(req, res){
-  const { id } = req.params
+  const { id } = req.params;
+
+  if (!isValidUUID) {
+    return res.status(400).json({ error: 'Contact not found'});
+  }
 
   const contact = await ContactsRepository.findById(id);
+
   if (!contact) {
     return res.status(404).json({ error: 'Contact not found' });
   }
@@ -22,17 +28,26 @@ class ContactController {
  async store(req, res){
   const { name, email, phone, category_id } = req.body;
 
+  if (category_id && !isValidUUID(category_id)) {
+    return res.status(400).json({ error: 'Invalid category'});
+  }
+
   if (!name) {
     return res.status(400).json({ error: 'Name is required' });
   }
 
-  const contactExists = await ContactsRepository.findByEmail(email);
-  if (contactExists){
-    return res.status(400).json({error: 'This e-mail is already in use '});
+  if (email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
+      if (contactExists){
+        return res.status(400).json({error: 'This e-mail is already in use '});
+      }
   }
 
   const contact = await ContactsRepository.create({
-    name, email, phone, category_id,
+    name,
+    email: email || null,
+    phone,
+    category_id: category_id || null,
   });
 
   res.status(201).json(contact);
@@ -42,21 +57,35 @@ class ContactController {
   const { id } = req.params;
   const { name, email, phone, category_id } = req.body;
 
-  const contactExists = await ContactsRepository.findById(id);
-  if (!contactExists) {
-    return res.status(404).json({ error: 'Contact not found' });
+  if (!isValidUUID) {
+    return res.status(400).json({ error: 'Contact not found'});
+  }
+
+  if (category_id && !isValidUUID(category_id)) {
+    return res.status(400).json({ error: 'Invalid category'});
   }
 
   if (!name) {
     return res.status(400).json({ error: 'Name is required' });
   }
 
-  const contactByEmail = await ContactsRepository.findByEmail(email);
-  if (contactByEmail && contactByEmail.id !== id) {
-    return res.status(400).json({error: 'This e-mail is already in use'});
+  const contactExists = await ContactsRepository.findById(id);
+  if (!contactExists) {
+    return res.status(404).json({ error: 'Contact not found' });
   }
+
+  if (email) {
+    const contactByEmail = await ContactsRepository.findByEmail(email);
+      if (contactByEmail && contactByEmail.id !== id) {
+        return res.status(400).json({error: 'This e-mail is already in use'});
+      }
+  }
+
   const contact = await ContactsRepository.update(id, {
-    name, email, phone, category_id
+    name,
+    email: email || null,
+    phone,
+    category_id: category_id || null ,
   });
 
   res.json(contact);
@@ -64,6 +93,10 @@ class ContactController {
 
 async delete(req, res){
   const { id } = req.params;
+
+  if (!isValidUUID) {
+    return res.status(400).json({ error: 'Contact not found'});
+  }
 
   await ContactsRepository.delete(id);
   res.sendStatus(204);
